+++ modules/mod_joy/mod_joy.c
@@ -1,7 +1,7 @@
 /*
- *  Copyright © 2006-2013 SplinterGU (Fenix/Bennugd)
- *  Copyright © 2002-2006 Fenix Team (Fenix)
- *  Copyright © 1999-2002 José Luis Cebrián Pagüe (Fenix)
+ *  Copyright Â© 2006-2013 SplinterGU (Fenix/Bennugd)
+ *  Copyright Â© 2002-2006 Fenix Team (Fenix)
+ *  Copyright Â© 1999-2002 JosÃ© Luis CebriÃ¡n PagÃ¼e (Fenix)
  *
  *  This file is part of Bennu - Game Development
  *
@@ -53,7 +53,7 @@
  *
  **/
 
-static int modjoy_num( INSTANCE * my, int * params )
+static int modjoy_num( INSTANCE * my, intptr_t * params )
 {
     return ( libjoy_num() );
 }
@@ -66,7 +66,7 @@ static int modjoy_num( INSTANCE * my, int * params )
  *
  **/
 
-static int modjoy_name( INSTANCE * my, int * params )
+static int modjoy_name( INSTANCE * my, intptr_t * params )
 {
     return ( libjoy_name( params[0] ) );
 }
@@ -80,7 +80,7 @@ static int modjoy_name( INSTANCE * my, int * params )
  *
  **/
 
-static int modjoy_select( INSTANCE * my, int * params )
+static int modjoy_select( INSTANCE * my, intptr_t * params )
 {
     return ( libjoy_select( params[0] ) );
 }
@@ -94,7 +94,7 @@ static int modjoy_select( INSTANCE * my, int * params )
  *
  **/
 
-static int modjoy_buttons( INSTANCE * my, int * params )
+static int modjoy_buttons( INSTANCE * my, intptr_t * params )
 {
     return ( libjoy_buttons() ) ;
 }
@@ -108,7 +108,7 @@ static int modjoy_buttons( INSTANCE * my, int * params )
  *
  **/
 
-static int modjoy_axes( INSTANCE * my, int * params )
+static int modjoy_axes( INSTANCE * my, intptr_t * params )
 {
     return ( libjoy_axes() ) ;
 }
@@ -122,7 +122,7 @@ static int modjoy_axes( INSTANCE * my, int * params )
  *
  **/
 
-static int modjoy_get_button( INSTANCE * my, int * params )
+static int modjoy_get_button( INSTANCE * my, intptr_t * params )
 {
     return ( libjoy_get_button( params[0] ) ) ;
 }
@@ -137,7 +137,7 @@ static int modjoy_get_button( INSTANCE * my, int * params )
  *
  **/
 
-static int modjoy_get_position( INSTANCE * my, int * params )
+static int modjoy_get_position( INSTANCE * my, intptr_t * params )
 {
     return ( libjoy_get_position( params[0] ) ) ;
 }
@@ -150,7 +150,7 @@ static int modjoy_get_position( INSTANCE * my, int * params )
  *
  **/
 
-static int modjoy_hats( INSTANCE * my, int * params )
+static int modjoy_hats( INSTANCE * my, intptr_t * params )
 {
     return ( libjoy_hats() ) ;
 }
@@ -163,7 +163,7 @@ static int modjoy_hats( INSTANCE * my, int * params )
  *
  **/
 
-static int modjoy_balls( INSTANCE * my, int * params )
+static int modjoy_balls( INSTANCE * my, intptr_t * params )
 {
     return ( libjoy_balls() ) ;
 }
@@ -176,7 +176,7 @@ static int modjoy_balls( INSTANCE * my, int * params )
  *
  **/
 
-static int modjoy_get_hat( INSTANCE * my, int * params )
+static int modjoy_get_hat( INSTANCE * my, intptr_t * params )
 {
     return ( libjoy_get_hat( params[0] ) ) ;
 }
@@ -189,14 +189,14 @@ static int modjoy_get_hat( INSTANCE * my, int * params )
  *
  **/
 
-static int modjoy_get_ball( INSTANCE * my, int * params )
+static int modjoy_get_ball( INSTANCE * my, intptr_t * params )
 {
     return ( libjoy_get_ball( params[0], ( int* )params[1], ( int* )params[2] ) ) ;
 }
 
 /* --------------------------------------------------------------------------- */
 
-static int modjoy_get_accel( INSTANCE * my, int * params )
+static int modjoy_get_accel( INSTANCE * my, intptr_t * params )
 {
     return ( libjoy_get_accel( ( int * ) params[0], ( int * ) params[1], ( int * ) params[2] ) );
 }
@@ -214,7 +214,7 @@ static int modjoy_get_accel( INSTANCE * my, int * params )
  *
  **/
 
-static int modjoy_buttons_specific( INSTANCE * my, int * params )
+static int modjoy_buttons_specific( INSTANCE * my, intptr_t * params )
 {
     return ( libjoy_buttons_specific( params[0] ) ) ;
 }
@@ -228,7 +228,7 @@ static int modjoy_buttons_specific( INSTANCE * my, int * params )
  *
  **/
 
-static int modjoy_axes_specific( INSTANCE * my, int * params )
+static int modjoy_axes_specific( INSTANCE * my, intptr_t * params )
 {
     return ( libjoy_axes_specific( params[0] ) ) ;
 }
@@ -242,7 +242,7 @@ static int modjoy_axes_specific( INSTANCE * my, int * params )
  *
  **/
 
-static int modjoy_get_button_specific( INSTANCE * my, int * params )
+static int modjoy_get_button_specific( INSTANCE * my, intptr_t * params )
 {
     return ( libjoy_get_button_specific( params[0], params[1] ) );
 }
@@ -257,7 +257,7 @@ static int modjoy_get_button_specific( INSTANCE * my, int * params )
  *
  **/
 
-static int modjoy_get_position_specific( INSTANCE * my, int * params )
+static int modjoy_get_position_specific( INSTANCE * my, intptr_t * params )
 {
     return ( libjoy_get_position_specific( params[0], params[1] ) );
 }
@@ -272,7 +272,7 @@ static int modjoy_get_position_specific( INSTANCE * my, int * params )
  *
  **/
 
-static int modjoy_hats_specific( INSTANCE * my, int * params )
+static int modjoy_hats_specific( INSTANCE * my, intptr_t * params )
 {
     return ( libjoy_hats_specific( params[0] ) ) ;
 }
@@ -285,7 +285,7 @@ static int modjoy_hats_specific( INSTANCE * my, int * params )
  *
  **/
 
-static int modjoy_balls_specific( INSTANCE * my, int * params )
+static int modjoy_balls_specific( INSTANCE * my, intptr_t * params )
 {
     return ( libjoy_balls_specific( params[0] ) ) ;
 }
@@ -298,7 +298,7 @@ static int modjoy_balls_specific( INSTANCE * my, int * params )
  *
  **/
 
-static int modjoy_get_hat_specific( INSTANCE * my, int * params )
+static int modjoy_get_hat_specific( INSTANCE * my, intptr_t * params )
 {
     return ( libjoy_get_hat_specific( params[0], params[1] ) );
 }
@@ -311,14 +311,14 @@ static int modjoy_get_hat_specific( INSTANCE * my, int * params )
  *
  **/
 
-static int modjoy_get_ball_specific( INSTANCE * my, int * params )
+static int modjoy_get_ball_specific( INSTANCE * my, intptr_t * params )
 {
     return ( libjoy_get_ball_specific( params[0], params[1], ( int * ) params[2], ( int * ) params[3] ) );
 }
 
 /* --------------------------------------------------------------------------- */
 
-static int modjoy_get_accel_specific( INSTANCE * my, int * params )
+static int modjoy_get_accel_specific( INSTANCE * my, intptr_t * params )
 {
     return ( libjoy_get_accel_specific( params[0], ( int * ) params[1], ( int * ) params[2], ( int * ) params[3] ) );
 }
