+++ modules/mod_say/mod_say.c
@@ -1,7 +1,7 @@
 /*
- *  Copyright © 2006-2013 SplinterGU (Fenix/Bennugd)
- *  Copyright © 2002-2006 Fenix Team (Fenix)
- *  Copyright © 1999-2002 José Luis Cebrián Pagüe (Fenix)
+ *  Copyright Â© 2006-2013 SplinterGU (Fenix/Bennugd)
+ *  Copyright Â© 2002-2006 Fenix Team (Fenix)
+ *  Copyright Â© 1999-2002 JosÃ© Luis CebriÃ¡n PagÃ¼e (Fenix)
  *
  *  This file is part of Bennu - Game Development
  *
@@ -35,7 +35,7 @@
 
 /* ---------------------------------------------------------------------- */
 
-static int modsay_say( INSTANCE * my, int * params )
+static int modsay_say( INSTANCE * my, intptr_t * params )
 {
     /* Show debugging info also in stdout */
     printf( "%s\n", string_get( params[0] ) );
@@ -46,7 +46,7 @@ static int modsay_say( INSTANCE * my, int * params )
 
 /* ---------------------------------------------------------------------- */
 
-static int modsay_say_fast( INSTANCE * my, int * params )
+static int modsay_say_fast( INSTANCE * my, intptr_t * params )
 {
     /* Show debugging info also in stdout */
     printf( "%s\n", string_get( params[0] ) );
